+++ Userland/Libraries/LibC/sys/sysmacros.h
@@ -6,7 +6,7 @@
 
 #pragma once
 
-#include <Kernel/API/Device.h>
+#include <sys/device.h>
 
 #define makedev(major, minor) serenity_dev_makedev((major), (minor))
 #define major(dev) serenity_dev_major(dev)
